backend/requirements.txt
# Backend dependencies for Phase II Todo Application
# Task: P2-T-003
# From: specs/phase-ii/plan

fastapi==0.115.0
uvicorn==0.32.0
sqlmodel==0.0.22
pydantic==2.10.3
python-multipart==0.0.20
python-dotenv==1.0.1
asyncpg==0.30.0
alembic==1.13.3
psycopg2-binary==2.9.10
httpx==0.27.2
pytest==8.3.4
pytest-cov==5.0.0
black==24.10.0
ruff==0.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0